% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{CoinTosses}
\alias{CoinTosses}
\title{Coin Tosses}
\usage{
CoinTosses(n, prob_heads = 0.5)
}
\description{
This function generates a sequence of \code{n} coin tosses with \code{prob_heads},
encoded as a character vector of length \code{n} containing only \code{'H'} and \code{'T'}.
}
\details{
The R way to do this is:
\verb{sample(c("H", "T"), n, replace = TRUE,}
\verb{ prob = c(prob_heads, 1 - prob_heads))}.
}
