% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{DieRolls}
\alias{DieRolls}
\title{Die Rolls}
\usage{
DieRolls(n)
}
\description{
Uses \code{RandomNumbers} directly to generate \code{n} six-sided die rolls.
The R way to do this is:
\code{sample(1:6, n, replace = TRUE)}
}
