% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{HTSimulation}
\alias{HTSimulation}
\title{Heads and Tails Simulation}
\usage{
HTSimulation(m, n, prob_heads = 0.5)
}
\description{
Simulates and summarizes the distribution of a trajectory of a trivial game
based on \code{n} coin tosses, interpreted as \code{H}eads winning $1 and \code{T} losing $1.
}
\details{
The distribution is estimated by sampling \code{m} trajectories.

Returns a \code{tibble} with two columns each of length \code{m}:
winnings: the total winnings at the end of the \code{n}th round.
ahead: the total number of rounds during which the player was ahead.
}
