% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{HTSimulationHex}
\alias{HTSimulationHex}
\title{Heads and Tail Simulation Hexplot}
\usage{
HTSimulationHex(hts)
}
\description{
Instead of SpikeGraph, visualize the outcome of \code{HTSimulation}
by using a hexplot. This requires a fairly large \code{m}.
}
