% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{HTSimulationHist}
\alias{HTSimulationHist}
\title{Heads and Tail Simulation Hexplot}
\usage{
HTSimulationHist(hts)
}
\description{
Instead of SpikeGraph, visualize the outcome of \code{HTSimulation}
by using two marginal histograms, one for each column.
}
