% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{HTSimulationTrace}
\alias{HTSimulationTrace}
\title{Heads and Tails Simulation Trace}
\usage{
HTSimulationTrace(m, n, prob_heads = 0.5)
}
\description{
Simulate and plot \code{m} trajectories of the coin toss game.
NOTE: this is run on its own, without input from \code{HTSimulation}.
}
