% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntroProbR.r
\docType{package}
\name{IntroProbR-package}
\alias{IntroProbR}
\alias{IntroProbR-package}
\title{IntroProbR: Implements "Programs" from _Introduction to Probability_ by Grinstead and Snell.}
\description{
R Implementations of "Programs" from \emph{Introduction to Probability} by Grinstead and Snell.
}
\author{
\strong{Maintainer}: Timothy Teräväinen \email{teravainen@gmail.com}

}
\keyword{internal}
