% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{MonteCarlo}
\alias{MonteCarlo}
\title{Monte Carlo}
\usage{
MonteCarlo(n, f = function(x) x^2, sample_y = TRUE)
}
\description{
This function implements Monte Carlo integration with \code{n} steps,
estimating the area of a function, \code{f}, under the unit square.
With \code{sample_y = TRUE}, the default, the method in the text is used.
With \code{sample_y = FALSE}, the x points are sampled but f(x) is directly evaluated.
}
