% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{SummarizeCoinTosses}
\alias{SummarizeCoinTosses}
\title{Summarize Coin Tosses}
\usage{
SummarizeCoinTosses(coin_tosses)
}
\description{
Generate a named vector of summary statistics of a coin toss vector (above).
}
